Restructure package layout for main and test sources
Grouped main classes into core and utils packages for better separation of concerns
