+++ src/main/kotlin/racingcar/utlls/CarFactoring.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.utlls
+
+import racingcar.core.Car
 
 object CarFactoring {
