+++ src/main/kotlin/racingcar/utlls/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.utlls
 
 import camp.nextstep.edu.missionutils.Console
