+++ src/test/kotlin/racingcar/utils/CarFactoringTest.kt
@@ -1,9 +1,10 @@
-package racingcar
+package racingcar.utils
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.utlls.CarFactoring
 
- class CarFactoringTest {
+class CarFactoringTest {
 
      @Test
      fun testCreateCars() {
