+++ src/test/kotlin/racingcar/utils/InputViewTest.kt
@@ -1,9 +1,10 @@
-package racingcar
+package racingcar.utils
 
 import org.junit.jupiter.api.Test
 
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.assertThrows
+import racingcar.utlls.InputView
 
 class InputViewTest {
